Move spotify search API call to own file
